+++ R/fix-panel.R
@@ -13,6 +13,11 @@
 
 
 fix_panel <- function(plot, width = NULL, height = NULL){
+
+  # check arguments
+  if (!missing(width) & !missing(height)){stop("Only specificy one of width and height. (Aspect ratio is maintained.)")}
+
+  # generate grob from ggplot2 plot
   gtable <- ggplot2::ggplotGrob(plot)
 
   ## calculate ratio from numeric part of panel height / width